requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
asttokens==3.0.1
    # via stack-data
catboost==1.2.8
    # via -r requirements.in
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   ipython
comm==0.2.3
    # via ipykernel
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.19
    # via ipykernel
decorator==5.2.1
    # via ipython
executing==2.2.1
    # via stack-data
fastapi==0.128.0
    # via -r requirements.in
fonttools==4.61.1
    # via matplotlib
graphviz==0.21
    # via catboost
h11==0.16.0
    # via uvicorn
idna==3.11
    # via anyio
ipykernel==7.1.0
    # via -r requirements.in
ipython==9.9.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via -r requirements.in
joblib==1.5.3
    # via
    #   -r requirements.in
    #   scikit-learn
jupyter-client==8.8.0
    # via ipykernel
jupyter-core==5.9.1
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.9
    # via matplotlib
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via
    #   -r requirements.in
    #   catboost
matplotlib-inline==0.2.1
    # via
    #   ipykernel
    #   ipython
narwhals==2.15.0
    # via plotly
nest-asyncio==1.6.0
    # via ipykernel
numpy==2.4.1
    # via
    #   -r requirements.in
    #   catboost
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
packaging==25.0
    # via
    #   ipykernel
    #   matplotlib
    #   plotly
pandas==2.3.3
    # via
    #   -r requirements.in
    #   catboost
parso==0.8.5
    # via jedi
pillow==12.1.0
    # via matplotlib
platformdirs==4.5.1
    # via jupyter-core
plotly==6.5.2
    # via catboost
prompt-toolkit==3.0.52
    # via ipython
psutil==7.2.1
    # via ipykernel
pure-eval==0.2.3
    # via stack-data
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pyparsing==3.3.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via -r requirements.in
python-multipart==0.0.21
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via -r requirements.in
pyzmq==27.1.0
    # via
    #   ipykernel
    #   jupyter-client
scikit-learn==1.8.0
    # via -r requirements.in
scipy==1.17.0
    # via
    #   -r requirements.in
    #   catboost
    #   scikit-learn
six==1.17.0
    # via
    #   catboost
    #   python-dateutil
stack-data==0.6.3
    # via ipython
starlette==0.50.0
    # via fastapi
threadpoolctl==3.6.0
    # via scikit-learn
tornado==6.5.4
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.14.3
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
uvicorn==0.40.0
    # via -r requirements.in
wcwidth==0.2.14
    # via prompt-toolkit
